pyproject.toml
[tool.poetry]
name = "space-stations"
version = "0.1.0"
description = ""
authors = ["Павел Сергеев <pavel.al.sergeev@gmail.com>"]
readme = "README.md"
packages = [{include = "space_stations"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
coverage = "^6.5.0"
flake8-broken-line = "^0.6.0"
flake8-isort = "^5.0.3"
flake8-plugin-utils = "^1.3.2"
flake8-return = "^1.2.0"
pep8-naming = "^0.13.2"
drf-spectacular = "^0.24.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
